skip the height and let the front end fake it
The reason I'm hacking it for now is that the steering output works in 3d and that
will give an output in the y axis that be almost nullified if we clear out the position of
moving object on every frame. This will only happen when object and targets are of widely
different height.